//CRM-3667 //delete case& related activities-batch update.
git-svn-id: http://svn.civicrm.org/civicrm/trunk@18439 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Activity/BAO/Activity.php
@@ -134,15 +134,20 @@ class CRM_Activity_BAO_Activity extends CRM_Activity_DAO_Activity
      * @access public
      *
      */
-    public function deleteActivity( &$params ) 
+    public function deleteActivity( &$params, $moveToTrash = false ) 
     {
         require_once 'CRM/Core/Transaction.php';
         $transaction = new CRM_Core_Transaction( );
-        
-        $activity     =& new CRM_Activity_DAO_Activity( );
+               
+        $activity    =& new CRM_Activity_DAO_Activity( );
         $activity->copyValues( $params );
-        $result = $activity->delete( );
         
+        if ( ! $moveToTrash ) {  
+            $result = $activity->delete( );
+        } else {
+            $activity->is_deleted = 1;
+            $result = $activity->save( );
+        }
         $transaction->commit( );
         return $result;
     }
